#145 [ADD] mainPage -> index로 페이지명 수정
